requirements/production.txt
requests==1.2.3
celery==3.0.15
redis==2.7.6
PyYAML==3.10
#graypy==0.2.9
graypy==0.2.7
msgpack-python==0.3.0
numpy==1.7.1
scipy==0.12.0
